+++ CHANGELOG.md
@@ -17,6 +17,8 @@ releases may include breaking changes.
 - ⬆️ Raise the macOS deployment target to 13.3 to enable `std::format` in libc++
   ([#976]) ([**@denialhaag**])
 - 💥 Require Python 3.11 or newer ([#976]) ([**@denialhaag**])
+- 🚚 Move recursive unitary construction from MQT Core to MQT DDSIM ([#975])
+  ([**@simon1hofmann**])
 
 ## [2.5.0] - 2026-08-20
 
@@ -173,6 +175,7 @@ _📚 Refer to the [GitHub Release Notes] for previous changelogs._
 <!-- PR links -->
 
 [#976]: https://github.com/munich-quantum-toolkit/ddsim/pull/976
+[#975]: https://github.com/munich-quantum-toolkit/ddsim/pull/975
 [#966]: https://github.com/munich-quantum-toolkit/ddsim/pull/966
 [#955]: https://github.com/munich-quantum-toolkit/ddsim/pull/955
 [#954]: https://github.com/munich-quantum-toolkit/ddsim/pull/954